==> 5_cluster_representatives.Rmd <==
Step 5. Use cluster representative sequences
==============================================

**Author**: Philipp Muench (pmu15@helmholtz-hzi.de)

**Date**: `r Sys.Date()`

----------------------------------------

<style type="text/css">
h1 {
    line-height: 120%;
}
</style>


load libraries 

```{r, eval=F}
library(tidyverse)
library(ggtree)
```

plot the tree of the representative sequences 

```{r, fig.width=9, fig.height=7}
tree <- read.tree("cluster_rep/colicin.tree")
ggtree(tree) + geom_tiplab() + ggplot2::xlim(0, 0.51) 
```

we can process the downloaded genomes in the FASTA format using [diamond](https://github.com/bbuchfink/diamond) blastp. After installing diamond to the `tools` folder we can create a database using

```{bash, eval=F}
tools/diamond makedb --in ../cluster_rep/colicin.fasta -d ../diamond_colicin_database
```

to screen one fasta file 

```{bash, eval=F}
tools/diamond blastp -d diamond_colicin_database.dmnd -q genomes/faa/test.faa -o genomes/faa/test_matches.m8
```

to screen all fasta files we can run
ubject-cover
```{bash, eval=F}
parallel 'tools/diamond blastp -d diamond_colicin_database.dmnd --more-sensitive --id 55 -p 1 -q {} -o {.}.matches.m8 {}' ::: genomes/faa/*.faa
```

and we can submit such job on the cluster using `qsub -pe smp 4 -l h_vmem=4G diamond.sh`
